requirements.txt
streamlit>=1.28
anthropic>=0.28
python-dotenv>=1.0
